# backend/swagger.yaml
openapi: 3.0.3
info:
  title: ERP System Backend API
  description: |
    Complete API documentation for the ERP System Backend microservices.
    
    ## Services
    - **Auth Service**: Authentication and user management
    - **Organization Service**: Organizations, companies, and locations
    - **Inventory Service**: Products and stock management
    - **License Service**: Application licenses and assignments
    - **Subscription Service**: Subscription plans and billing
    
    ## Authentication
    Most endpoints require JWT authentication via Bearer token in the Authorization header.
    
  version: 1.0.0
  contact:
    name: ERP System Team
    email: support@erpsystem.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8081/api/v1
    description: Auth Service (Development)
  - url: http://localhost:8082/api/v1
    description: Organization Service (Development)
  - url: http://localhost:8083/api/v1
    description: Inventory Service (Development)
  - url: http://localhost:8084/api/v1
    description: License Service (Development)
  - url: http://localhost:8085/api/v1
    description: Subscription Service (Development)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or register endpoint
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        code:
          type: integer
          example: 200
        message:
          type: string
        data:
          type: object
...
    BaseModel:
      type: object
      properties:
        id:
          type: string
          format: objectid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        created_by:
          type: string
          format: objectid
        updated_by:
          type: string
          format: objectid
          nullable: true
        deleted_at:
          type: string
          format: date-time
          nullable: true
        deleted_by:
          type: string
          format: objectid
          nullable: true
...
    Product:
      allOf:
        - $ref: '#/components/schemas/BaseModel'
        - type: object
          properties:
            organization_id:
              type: string
              format: objectid
            name:
              type: string
            sku:
              type: string
            barcode:
              type: string
            description:
              type: string
            category_id:
              type: string
              format: objectid
            unit_of_measure:
              type: string
            status:
              type: string
              enum: [active, inactive, discontinued]
            track_inventory:
              type: boolean
            reorder_level:
              type: integer
            reorder_quantity:
              type: integer
    CreateProductRequest:
      type: object
      required:
        - name
        - sku
      properties:
        name:
          type: string
        sku:
          type: string
        barcode:
          type: string
        description:
          type: string
        category_id:
          type: string
        unit_of_measure:
          type: string
        track_inventory:
          type: boolean
        reorder_level:
          type: integer
        reorder_quantity:
          type: integer
    
    # License Schemas
...
paths:
...
  /organizations/{org_id}/products:
    get:
      tags:
        - Inventory
      summary: List products
      description: Get paginated list of products with optional filters
      security:
        - BearerAuth: []
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Products retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
    
    post:
      tags:
        - Inventory
      summary: Create product
      description: Create a new product
      security:
        - BearerAuth: []
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Product'
  /products/{id}:
    get:
      tags:
        - Inventory
      summary: Get product
      description: Retrieve product by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Product'
    
    put:
      tags:
        - Inventory
      summary: Update product
      description: Update existing product
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Product'
    
    delete:
      tags:
        - Inventory
      summary: Delete product
      description: Soft delete a product
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...